* removed some old Qt3 constructors * moved the property system from uint to int
svn path=/trunk/KDE/kdeedu/kig/; revision=464489

+++ modes/label.cc
@@ -140,7 +140,7 @@ void TextLabelModeBase::leftReleased( QMouseEvent* e, KigWidget* v )
     QAction* act = p->addAction( i18n( "Name" ) );
     act->setData( QVariant::fromValue( 0 ) );
     QByteArrayList l = o->imp()->properties();
-    assert( static_cast<uint>( l.size() ) == o->imp()->numberOfProperties() );
+    assert( l.size() == o->imp()->numberOfProperties() );
     for ( int i = 0; i < l.size(); ++i )
     {
       QString s = i18n( l[i] );
